phase-2/backend/requirements.txt
fastapi==0.109.0
sqlmodel==0.0.14
pyjwt[crypto]==2.8.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.26.0
